lesson 8 Arithmetic type change int -> long

+++ src/lesson8/accounts/arithmetic/Adder.java
@@ -1,7 +1,7 @@
 package lesson8.accounts.arithmetic;
 
 public class Adder extends Arithmetic {
-    public int add(int a, int b){
+    public long add(int a, int b){
         return a + b;
     }
 }
